+++ src/main/java/com/shy_polarbear/server/domain/user/controller/AuthController.java
@@ -1,11 +1,11 @@
 package com.shy_polarbear.server.domain.user.controller;
 
-import com.shy_polarbear.server.domain.user.dto.auth.LogoutResponse;
-import com.shy_polarbear.server.domain.user.dto.auth.ReissueRequest;
+import com.shy_polarbear.server.domain.user.dto.auth.response.LogoutResponse;
+import com.shy_polarbear.server.domain.user.dto.auth.request.ReissueRequest;
 import com.shy_polarbear.server.global.common.dto.ApiResponse;
 import com.shy_polarbear.server.global.auth.jwt.JwtDto;
-import com.shy_polarbear.server.domain.user.dto.auth.JoinRequest;
-import com.shy_polarbear.server.domain.user.dto.auth.SocialLoginRequest;
+import com.shy_polarbear.server.domain.user.dto.auth.request.JoinRequest;
+import com.shy_polarbear.server.domain.user.dto.auth.request.SocialLoginRequest;
 import com.shy_polarbear.server.domain.user.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
