+++ src/main/java/com/shy_polarbear/server/domain/user/dto/user/response/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.user.dto.user;
+package com.shy_polarbear.server.domain.user.dto.user.response;
 
 import com.shy_polarbear.server.domain.user.model.User;
 import lombok.Builder;
